requirements.txt
aiocontextvars==0.2.2
allure-pytest==2.8.6
allure-python-commons==2.8.6
apipkg==1.5
appdirs==1.4.4
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.1.0
contextvars==2.4
cryptography==2.9.2
dataclasses==0.7
execnet==1.7.1
filetype==1.0.7
har2case==0.3.1
HTMLParser==0.0.2
HttpRunner==1.5.8
idna==2.8
immutables==0.14
importlib-metadata==1.6.0
Jinja2==2.11.2
jmespath==0.9.5
loguru==0.4.1
lxml==4.5.0
MarkupSafe==1.1.1
more-itertools==8.2.0
numpy==1.17.2
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pycryptodome==3.9.7
pydantic==1.5.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
pytest-forked==1.1.3
pytest-repeat==0.7.0
pytest-rerunfailures==9.0
pytest-xdist==1.23.2
PyYAML==3.12
regex==2020.5.14
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.14.1
six==1.14.0
soupsieve==1.9.5
toml==0.10.1
typed-ast==1.4.1
urllib3==1.24
wcwidth==0.1.9
win32-setctime==1.0.1
zipp==3.1.0
